Refactor: Separate data/record types, extract FinanceAuthService, add delete method
