+++ client/finance/services/auth/AuthorizationService.ts
@@ -1,38 +1,12 @@
-import AuthService from '@common/services/auth/AuthService';
-
 import AuthUtil from '@client-common/auth/AuthUtil';
 
-import { FinanceAuthDataType } from '@/interfaces/data/FinanceAuthDataType';
-import { FinanceAuthRecordType } from '@/interfaces/records/FinanceAuthRecordType';
 import PermissionMatrixService from '@/services/auth/PermissionMatrixService';
+import FinanceAuthService from '@/services/auth/FinanceAuthService';
 import {
   Feature,
   PermissionLevel,
   UserType,
 } from '@/types/AuthorizationTypes';
-
-/**
- * Finance用の認証サービス
- */
-class FinanceAuthService extends AuthService<FinanceAuthDataType, FinanceAuthRecordType> {
-  public constructor() {
-    super(FinanceAuthService.dataToRecord, FinanceAuthService.recordToData);
-  }
-
-  private static dataToRecord(data: FinanceAuthDataType): FinanceAuthRecordType {
-    return {
-      ...AuthService.dataToRecordBase(data),
-      Finance: data.finance
-    };
-  }
-
-  private static recordToData(record: FinanceAuthRecordType): FinanceAuthDataType {
-    return {
-      ...AuthService.recordToDataBase(record),
-      finance: record.Finance
-    };
-  }
-}
 
 /**
  * 汎用認可サービス
